Implement would not take config for first use
If the Manage Impements screen was empty, and an implement like a Tome was added, the initial rules data would not get added to that implement because the name of the implement is drawn from the existing implement. If the existing implement does not exist, use the name of the newly selected one instead.