Implement helper to get implement data
This will find the implement data in the flags and return it, or null if
the implement isn't one the actor has.  The implement is specified via a
slug-like name, e.g. "tome", or "weapon".  It's not for user display and
wouldn't be translated.

Currently it uses a kludge method of translating the slug into the
localized implement name, and then looking for a flag with a matching
label.  But this could change and then users of the helper would all
just keep working but with whatever faster and more reliable method gets
used.