Use getImplement for tome, amulet
Somewhat shorter.  Also, should be easier to change things with the
flags without having to edit as much code afterward.

This also removes the check for having the Thaumaturge class from the
Tome dialog.  It's faster to check for the implement flags than to check
for the class, and a non-thaum wouldn't have the implement flags.

And if the actor isn't a thaum, but has the flags, then they must be
some kind of archetype or homebrew thing that gets a thaum implement's
abilities.  And so giving the implement benefits would be the correct
thing to do.