Send result of RK check back to user who made it
The GM actually does the roll, so the user who made the request didn't
get the result.

Use a socket function to send the result back to them.  This means we
need also need to tell createRKDialog() the user, not just the actor,
who did the roll.

The callback doesn't do anything, but things that trigger on RK checks
could put hooks in there, like Tome Adept benefit.