Automate the Tome Adept RK +1 to attack benefit
This hooks on the RK check result to apply an effect for the RK benefit.
The effect uses a TokenMark rule to mark the target of the RK check.
The FlatModifier remains on the Tome item like it is now, but uses the
mark in the predicate to trigger on the right attack.  The hook code
will automatically set the TokenMark's uuid to the RK check's target's
uuid.

The effect expires at the end of the turn.  If an expired effect is
present, the RK will replace it with a new one.  An unexpired check
means this isn't the first RK in the turn and the effect isn't updated.

It's not automated yet that the effect goes away after an attack against
the RK target.  That needs to hook on attack rolls.

If someone wants the Tome Adept RK attack bonus, but didn't do the RK
roll with the macro, they can just drag the effect onto their actor and
it'll mark the current target or let them pick one.

If the first RK check of the turn is not acually the Tome Adept free
check, e.g. they skipped it, then decided to do a check mid-turn, then
the bonus shouldn't apply.  It'll get added anyway.  Just right-click
the effect to make it go away.

I thought about adding a dialog the RK macro, "Is this the Tome Adept
free check?" But that's more clicks to use that just removing the effect
in the rare case it's wrong.