Change amulet dialog width to fixed value.
The value used was based on the width of the full map size in pixels.
Not the size of the map viewport, but the entire virutal map used for
scene.  It isn't related to a dialog size and is generally way too big.

The character sheet is 750 pixels, so try that as a decent size.