Let system expire the amulet abeyance lingering effect
If the effect is created with context.origin.actor set (which the
createEffectData() helper will do), then it will expire in the origin
actor's turn, rather than the turn of the actor the effect is on.

With the effect's duration now set to 1 turn, it expires at the right
time.  There's already automation to expire and then remove effects.  So
the module hook can just go away.

Also create both effects in one call, and don't await the result, so it
goes faster when a bunch of effects are made at once.