Use preCreateChatMessage rather than createChatMessage
The former is only called once, for the client who creates the message,
before it's acutally done, and has the option to disable creating the
message by returning flase.

The latter is called once when the message is made, after it's done, can't
prevent it from getting created, and for every client.  I didn't notice that
last difference, and the rest seem like a better choice.