Newer checks for Amulet presence
Check the reaction module setting before we even look for thaum actors.

Check if an actor has an amulet by looking for
`actor.attributes.implements.amulet`.  This avoids the localized name
check of the class, which is problematic for dual class, archetypes,
etc.

It's easy enough to check `actor.isOwner` in the filter too.

The check inside the for loop has nothing left at this point.