Send abeyance effect data by id, not name
It's possible for tokens to have the same name.  Have the abeyance data
sent to generate the effect(s) use the token uuid as the property names
instead of the actor names.

There's no need to await a uuid lookup anymore and the callback doesn't
need to be async.

Also put the name of the thaumaturge creating the effect into the effect
name instead of the actor the effect is going on.

For the lingering resistence effect, append the damage type to the name.

Stop adding a pf2e-thaum-vuln.effectSource flag to the effect, since the
pf2e system already has that information in the effect origin and
nothing used that flag.