Use TokenMark exploit-vulnerability in effects
This fixes Breached Defenses not working.

Use the same mark name for PA, MW, and BD.  Pretty much all the rules
say "target of your EV" and don't care which type of EV was picked.
This will likely avoid needing to "or" the three types of EV effect in
every rule that uses it.

Change Esoteric Warden to use the mark.  This means the predicates in
the rule can always be "origin:mark:exploit-vulnerability" and the
modules doesn't need to change them.