Add roll options "self:implement:{slug}:rank:X" in derived data
This will be handy in various REs, for things like "adept or better in
tome" with {"gte": ["self:implement:tome:rank", 2}.

These options are added in the derived data wrapper.  So they aren't stored
in the database and so there's no need to keep them up to date as things on
the character change.  The system does this same thing with all the
"self:trait:*" roll options.

The getter for rank means "@actor.attributes.implements.regalia.rank" can
also be used in RE values for calculations, like Regalia aura radius.