Changes for Lantern and Intensify effects.
Standardize on the slug "intensify-vulnerability-{slug}" for the effects of
an intensified implement.  This will give the roll option
"self:effect:intensify-vulnerability-{slug}" to be checked in rule
elements.  The existing effects were not consistent on this.

Standardize on the name "Effect: Intensify Vulnerability (Implement)" for
the effects.  This is most like how other effects that come in different
versions are named.

Automate some more of the Lantern effects:

Use an AELike that calculates the radius of the light, saves it to an actor
flag.
Use just one Aura and one TokenLight, with the calculated radius, for the
three different sizes.

The Aura RE will push new effects to allies and enemies in the aura.
Allies will get "in-token-light-ally" while enemies get
"in-token-light-enemy".

This way "target:effect:in-token-light-enemy" can be used to determine if a
Lantern bonus should apply to the roll based on the target being in the
light.

The ally effect gives them the +1 bonus to RK and Perception.  The enemy
effect has a -2 to Stealth and Deception.

The IV effect has two AdjustModifier RE to change the value for the thaum's
bonus from +1 to +2, but only for the target marked by EV.  It also
relabels the bonuses to "… (Intensified)".

The selector for the RK bonus is changed from "all" to "skill-check".  This
way it doesn't show for attack rolls.  I believe all RK rolls will always
be skill checks.

The bonsues are labelled "In Lantern Light".  This fits better, and no
effects exist for the dim light, so bright is unnecessary and what the roll
is for is already known.

The perception bonus is no longer set to hide if disabled.  There's no way
to automate detecting if non-creature targets are in the light.

Give all effects a sourceId that matches their id.  The pf2e system does
this with it's compendiums.

There are few things which don't work yet:

Allies don't get a +2 bonus when IV is used and the roll is against the EV
target.  Getting the effect to know about IV, and which creature is the EV
target, is harder.

The -2 penalty for enemies when IV is used will still be off by default.
Same problem getting the effect to know if IV is on.

There's no way to automate the invisible/concealed stuff.

If there is more than one thaumaturge, it's possible for an ally in one
lantern's light to get a bonus against a target in another lantern's light.
There's no way to tell the effects apart.