Add attributes.implements.slug.intensified
This will indicate if the given implement has been intensified.  The
"intensify" property already there means they can be intensified, but not
that IV is active.

There is already a roll option from the IV effect, but there is no way to
query roll options with the "{actor|...}" or "@actor.attributes.…" syntax.
So I think this will be useful to get IV status places other than
predicates.

It should also work that an effect pushed to another actor by a thaum Aura
will be able to check "{item|origin.attributes.implements.…}" to get stuff
about the thaum's implements.  It's not possible for an effect from an aura
to check for roll options from the aura's actor.