Move effect creation into base Implement class
It will use a static class property with the Uuid of an effect, the
derived classes define it, and put that onto the actor.  If one isn't
defined, then it gives the unsupported error.

amulet, lantern, regalia, and weapon don't need intensifyImplement()
methods anymore.

The base method can also take optional effect data, tome uses this, since
it needs to modify the effect.

The check for implement being held is consistently done now.

Could also check for an existing IV effect here and remove old ones or
refuse to run.

The free function intensifyImplement() can just use the actor's
attributes.implement objects that are already there.  It doesn't need to
make a new one.  When the macro calls it, the macro can pass the `actor`
value, macros already get that, so it doesn't need to figure that out
itself.

Don't default to acting on game.user.character first, as the player can
control more than one character.  If they want to control a character other
than main one, they'll never be able to select it.

It's not necessary to specify the actor as the origin in an effect when the
effect is applied to that same actor.  That's the default.  It's only
needed when the effect is from someone else.

Change amulet IV effect predicate to use token mark

Use "origin:mark:exploit-vulnerability", instead of target, since ac is
a passive selector.

This way the rules don't need to be edited.