New code for checking if actor is Thaumaturge
Instead of splitting the class name into words and looking for translated
strings, check for the roll option.

Roll options are in the actor as a Map (we know exactly where to look and don't
need to do the full getRollOptions() gathering), so this is a fast O(1) lookup.
The dual class macro sets the roll options for both classes, so it works with
that too.

Put it into a helper, which will probably need to be tweaked for archetype
support.

Also, the hook isn't async.  It never awaits anything.