Cleanup for DragDrop setup on manageImplements dialog
Don't set dragSelector.  It would be used if the manage implements dialog had
drag sources on it.  But it doesn't, things can only be dragged to it, not from
it.

Don't put ".item-content-wrapper" dropSelector.  It's inside ".dropbox" div, so there's
no need to list it too.  The whole div will be a target.

It's not necessary to individually search for each drop target in the dialog and
bind them one at a time.  The dropSelector selects which element(s) become
targets, so bind can just be called once on the whole dialog and all the
elements matching the selector become drop targets.

We also don't want to search for drop targets starting at the `document` root,
as that searches the entire app.  It could find some other dialog's elements if
someone else used ".dropbox" as a class anywhere.  And it's inefficient to
search through hundreds of chat messages for dropbox element when we know we
only want the manage implements dialog.