Use checkImplements to setup actor flags
Find the implements by looking for feats with the thaumaturge otherTag.  Which
is how the system finds them for the choicesets.  And it's simpler.

Have checkImplements setup the selectedImplements flag.  It will have one entry
for each implement, with a uuid property that is null if an item wasn't selected
yet.  That's it, don't set all the other stuff that used to be there, like adept
ability or translated name, etc.

It'll remove implements from the flags that aren't part of the actor's feats and
null out an uuids that aren't in the inventory anymore.

Give checkImplements() a "clear" option to make a empty set of flags.  These
have the correct set of implements, but with all null uuids.

Await the setFlag() call and checkImplements(), so that attributes.implements is
up-to-date with the flags before the code continues.