Rework manage implements dialog
Big change here is to construct the data for the dialog hbs by using
actor.attributes.implements instead of a bunch of code using the actor flags and
looking for feats and so on.  This saves a ton of code.  A previous commit made
it so that checkImplements() will ensure the attributes are up-to-date, even if
the actor hasn't managed implements before.

Instead of passing data to the hbs in multiple objects, some indexed by slug and
some by translated implement name, just use one object with slug index.  Makes
the hbs templates a lot simpler.  The hbs "element" helper is now unused.

Get rid of a bunch of unused attributes on the html elements.  Stop using
non-standard html attributes for data like item uuid and use the element's
dataset.