Clean up items when generating packs
Do some checks on the documents in the packs.  The most important is to
make sure the _stats.compendiumSource value is correct.  Stuff breaks
when it's not.  Foundry doesn't validate the value when it loads a pack
or imports an item from a pack.

Also gets rid of user ID values in ownership and lastModifiedBy, since
the user won't exist on the server using the module.