Add two effects that give an EV DC SpecialStatistic
The pf2e system has an issue where you can't have a target for a check unless
that check is against a *named* DC on the target.  A DC specified numerically
doesn't work.  EV uses the standard level based DC for the target, which doesn't
exist as a named DC, so this is a problem.

It is possible to create a new named DC using the SpecialStatistic RE, and
that's what these effects do.  An EphemeralEffect RE on the roller can be used
to give these effects to the target of the EV check for the roll.  The new
statistic is named DC and targetting the check works.

There are two, one for the proficiency without level variant and one for normal
rules.  The system doesn't automatically subtract actor level from
SpecialStatistic proficiency bonuses when PwoL is used.