Give EV DC EphemeralEffect to thaums
This adds a synthetic rule during the derived data wrapper to give an
EphemeralEffect RE to thaum actors (with the EV feature).

The EE will give targets of EV a new DC to roll against for the EV check.

It would make sense to put this RE on the EV class feature, but that's a
system item and auto-modifiying those has issues.  Also, by doing it in
code, it can choose the normal of PwoL variant of the DC to apply.

It wouldn't work to put it on (each) implement, as the module doesn't
enforce holding an implement to use EV.