Convery EV check roll into an action
This reuses the SingleCheckAction class from the system for the mechanics of the
roll.  More of the system automation works this way.

The only thing it needs to do is setup the Diverse Lore DC text in the roll
note.  Maybe it's possible to get the system to inject that too?