Support rerolls for EV check
Hero point, keep higher, etc.

Uses the preCreateChatMessage hook to detect the roll result chat message when
the new message is posted.  The message will have isReroll set and the check DC
of "exploit-vulnerability" to identify it as an EV reroll.

Since preCreateChatMessage will only get called for the user who clicked the
reroll, we know we can handle the action and don't need to worry about
primaryUpdater or checking actor permissions.  We don't even need to check if
they are a thaum, since this is reroll they must be in order to have make the
original roll in the first place.

Refactor out the "delete old effects" and "create effect part" of the original
function to be called on reroll.

Deleting the old EW effects can be simplified some to use effect.origin to
determine if they are our EW effect or some another thaum.