Clean up for deleting EV effects
Simplify the logic for finding and deleting effects when EV changes.  Use the
system standard `origin` property to check if an effect was caused by the actor
whose effects are being deleted.

Fix crashes when there is a token with no actor in the scene.

Fixes Cursed Effigy not getting removed when EV is rerolled.

Fixes hero point re-rolling a critically failed EV check not removing the
Off-Guard effect.

Remove all the effects on the thaum actor in a single deleteEmbeddedDocuments
call, instead of using executeAsGM to have the GM's client remove them one at a
time.  This saves some round trips and makes doing a new EV roll faster.

Delete the targetID === "*" special mode of getActorEVEffect(), which is never
used.  Then delete TargetEffectSourceIDs which are no longer used.