Twin Weakness should only apply to the EV target
Sympathetic Vulnerabilities doesn't make the additional creatures that will
take extra damage targets of Exploit Vulnerability.  So things like
Intensify Vulnerability, Implement's Interruption, or Twin Weakness still
only work on the primary EV target.

All the rest were like this, but Twin Weakness allowed the additional SV
targets to be used.

After doing this, the flag EVTargetID is never used.  So the code to set it
can all be deleted too.

Also use the callback of the attack roll to get the DoS and trigger the
EXtra damage roll.  Rather than add and remove a hook and hope the next
attack roll message is the right one.