Add buttons to apply tome adept RK bonus to other RK rolls
The tome adept function worked for the esoteric lore RK macro in the
module, but not for rolls made using the workbench macro or the system
action.

This adds buttons to those rolls so after deciding if the result is a
success or failure the GM can click the button to apply the correct tome
adept RK effect.