Only send regalia reminders if at least one actor is frightened
Currently, this whisper list for the regalia always includes the GM, even if no one is frightened.

This reduces the incidence of whispers to when there is at least one frightened creature who is affected by the regalia.  This means the message will appear when it is more likely to have a real effect.