Get rid of last EffectOrigin custom flag
This flag is only still used in one place: to get which thaum created an PA or
MW effect.  It's not always the same player the effect is on because of Shared
Weakness.

But the system now has `effect.origin` that provides this and the custom flag
isn't needed.

Fix a place in the ForceEV macro where effects weren't generated using the
helper function, which means origin wouldn't be set on them.