Use custom weakness for PA
Now that they can be applyOnce weaknesses, it's possible to use it for PA.

There is still an effect on the target(s) of PA, but the weakness in that effect
is a custom one, with a definition of damage from the thaum.

That effect (and the other effects) will have the effect's level set to the
level of the thaum.  The value of the weakness can be set based on effect level.
So we don't need to modify the effect for that.

The hook that looked for strikes against a thaum target can stop looking for
targets with PA on them, since it doesn't need to patch the weaknesses type.

This doesn't affect MW, which is a bit more complex.